requirements.txt
structlog
Pillow
imgurpython
lxml
requests
PyYAML
discord
